% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fieldscorer.R
\name{Fieldscorer}
\alias{Fieldscorer}
\title{A package to convert Agrobase output files into Fieldscorer input files}
\usage{
Fieldscorer(x, Experiment, Trialtype, Barcode = NULL, Genotype)
}
\arguments{
\item{x}{A data.frame.}

\item{Experiment}{A character string indexing column that contains trial names.}

\item{Trialtype}{A character string to represent trial series. For example "21WS2E".}

\item{Genotype}{A character string indexing the column that contains genotype names.}

\item{plotbarcode}{A character string indexing column that contains plot barcodes.}

\item{Column}{A character string indexing the column that contains information pertaining to ranges. Fieldscorer accepts "Column" only.}

\item{Row}{A character string indexing the column that contains information pertaining to rows.}

\item{Pedigree}{A character string indexing the column that contains pedigree information.}
}
\value{
File format suitable for upload into the Fieldscorer app.
}
\description{
Fieldscorer takes an output file from Agrobase and converts it into a fieldscorer format.
}
\examples{
data <- read_xls("agrobasepractice.xls")

Fieldscorer(data, 
      Experiment = "trial.name",
      Trialtype = "21WS2E", 
      Barcode = "plot.barcode", 
      Column = "pcol", 
      Row = "prow",
      Genotype = "name",
      Pedigree = "pedigree")
}
\author{
Calum Watt, \email{cwatt@intergrain.com}
}
\keyword{utilities}
