% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVTconvert.R
\name{NVTconvert}
\alias{NVTconvert}
\title{A package to output NVT trial designs}
\usage{
NVTconvert(x, Crop.Species, Sub.Series = NULL, Location, State, Code)
}
\arguments{
\item{x}{A data.frame}

\item{Crop.Species}{User define crop species i.e. Wheat. Needs to be character input...."Wheat"}

\item{Sub.Series}{User defined crop sub-series. In wheat there are differne sub-series such as Durum. Default is NULL and will plot everything for a particular crop species.}

\item{Location}{A character stipulating the name of the column containing the location data...."Site.Location"}

\item{State}{A character stipulating the name of the column containing the state data...."State"}

\item{Code}{A character stipulating the name of the colunn containing the trialname data...."Trial.Name"}
}
\value{
Long format data frame and .jpeg trial design figures.
}
\description{
NVTconvert takes an NVT provided wide format trial designs and outputs each individual trial as a figure for a user define crop species.
}
\examples{
NVTconvert(data, 
      Crop.Species = "Wheat",
      Sub.Series = "Main Season", 
      Location = "SiteDescription", 
      State = "State", Code = "TrialCode")
}
\author{
Calum Watt, \email{cwatt@intergrain.com}
}
\keyword{utilities}
